backend/requirements.txt
fastapi==0.110.1
numpy==1.24.2
opencv_python==4.9.0.80
pandas==2.0.1
Pillow==10.3.0
scikit_learn==1.2.2
torch==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.66.2
transformers==4.39.3
